refactor: consolidate validation + move macros from direct.rs (#208, #209)
Issue #208: Extract duplicated validation logic (date format, symbol
format, interval, right, year) into a canonical `validate` module.
Both `endpoint.rs` and `direct.rs` now delegate to this single source
of truth. Build-time validators in `build_support/endpoints.rs` remain
separate (different domain) but include a cross-reference comment.

Issue #209: Move all macro_rules definitions (`list_endpoint!`,
`parsed_endpoint!`, `streaming_endpoint!`, and helper type-mapping
macros) from `direct.rs` to a dedicated `macros.rs` module. The module
is declared with `#[macro_use]` in `lib.rs` so the macros are visible
to both generated code and handwritten streaming endpoints.
